removed feature testing, refactored error messages on login and password reset, hashing passwords, all tests passing
