+++ .travis.yml
@@ -4,7 +4,3 @@ node_js:
 cache:
   directories:
     - 'node_modules'
-services:
-  - postgresql
-before_script:
-  - psql -c 'create database ketchup_test;' -U postgres
