+++ server/routes/forgot-password.js
@@ -3,16 +3,15 @@ const logIn = require('../lib/log-in')
 
 module.exports = {
   get: (_, res) => {
-    res.render('forgot-password.ejs', { flash: res.locals.flash })
+    res.render('forgot-password.ejs')
   },
   post: async (req, res) => {
     let result = await logIn.authenticate(req.body)
-    if (result.message === 'password incorrect') {
-      await forgotPassword.sendReset(req.body)
+    if (result.status) {
+      forgotPassword.sendReset(req.body)
       res.redirect('/')
     } else {
-      req.session.flash = { message: result.message }
-      res.redirect('/forgot-password')
+      res.redirect('/')
     }
   }
 }
