+++ server/routes/log-in.js
@@ -2,7 +2,7 @@ const logIn = require('../lib/log-in')
 
 module.exports = async (req, res) => {
   let result = await logIn.authenticate(req.body)
-  if (result.status) {
+  if (result.status && result.message === undefined) {
     req.session.user = result.user
     res.redirect('/dashboard')
   } else {
