--- test/migrations.js
@@ -1,3 +1,0 @@
-const run = require('../db/database-scripts')
-
-run.scripts('migrations', 'ketchup_test')
